Add beacon-shared-web-assets static content alias
This is a straightforward way for apache to serve the BSWA static content.
I didn't like the idea of adding another custom rule, but other options
did not prove to be better.

I looked for a way to use the same rule for common and shared. I didn't
come up with one.

I considered serving our content out of the static directory in target/
instead of src/ and target/. This would require a deploy for each file
change.

I considered unzipping BSWA content into the static directory in src/.
This may have caused file name conflicts, would have required a .gitignore
rule to prevent the shared content from being checked in, and developers
would be confused when changes in the dev src/ tree were did not appear
in git nor in non-dev.

I considered using a sym-link, which has many of the same problems as
unzipping content into src/.