Made tomcat-clean more effective, and called it less.
The tomcat-clean target originally only deleted the exploded webapp for the current
project, which led to it being called at deploy-time to avoid a situation where
tomcat might not notice that the exploded webapp file was out of date (since it
does not check for updated WARs at server startup).  Updated tomcat-clean to
delete any and all exploded directories that are older than their WAR files,
and removed it from the dependencies of the clean/deploy targets.