Tell apache to check tomcat availability on every request (for mac)
By default, if Apache can't connect to Tomcat, it'll display a
'Service Temporarily Unavailable' message for 60 seconds. It will
only check to see if tomcat is available after 60 seconds. This
commit has Apache check every time. This is particularly useful
when Tomcat is restarted, and while it is down, a site's keep-alive
triggers a request that tells Apache that Tomcat is down. Now,
Tomcat will show as available as soon as it is, instead of having
to wait for the retry period or restarting Apache.