Ivy repository cleaner: implemented necessary command-line options.
Added command-line option parsing to allow filtering revisions examined, actually doing the
deletions, deleting only non-matching files, and preserving certain revisions that would otherwise
have been removed (if, for example, this is a non-branched repo and we want to retain the builds
that are currently deployed to production).

Skipped pydoc, but added a large comment block to explain what those options (and the rest of the
script) are for.

Change-Id: Iadf6647655f5fcc86cca4bec1772fc131b5629b8