Ivy repository cleaner: reduced verbosity, increased brutality.
Added an option to list files that would be deleted, if no delete is being performed: previously
files were listed regardless, which was not always desired.  Relatively minimal output is now
feasible, if that's what you're looking for.

Also added an option to *not* always keep the last build of each patch-level, which is likely
to be most useful in cleaning the development/unmerged-jar branches.

Change-Id: I0ed49cd24d1d028d27f91196169b7c4f5e09dbd4