DRY fixes for properties, and suppressing spurious failures.
We have properties to say where the webapp source directory is and where the
freemarker directory is, but we also had the same information hard-coded in
various places: replaced the hard-codes with the appropriate properties.

We also have two recursive greps that fail if a given property is not set
(or the directory is not present): since the directory being missing probably
means that there are no files in it that violate the rule in question, we can
simply ignore the missing-directory error.

Change-Id: Ibd34e45165410dfb50f1a52298f9bd89a035746d