Include JSON files in jar files instead of WEB-INF.
For module publication, we want all resources to be published as a JAR file, with no
copying around during application build: the fixture JSON files were the only thing
that still relied on that kind of behavior.  Since wgspringcore will no longer be
publishing a web-assets zip file, that reliance needed to be broken.

Change-Id: Icea985e4b471824ccd3cf6737ac01b17cc2c1345