Actually run wgspringcore integration tests in all projects.
Some projects have stopped putting the wgspringcore integration-test jar
in "compile", since it is a test-time dependency for them, not a runtime
dependency.  Unfortunately, this means that those tests are not found by
the build script when it looks for wgspringcore integration tests to run.

Updating the script to look in ivy_lib/test (where these jars will always
be, as opposed to sometimes being in ivy_lib/compile), and also to look
for module integration tests as well as core integration tests.

Also updating the story build script, but wishing it weren't necessary.

Change-Id: I51c66884b95b5fb6aa91653da8d3846549a9f359