Prevent running destructive db targets in bad places
Make sure that the db host that targets are run against is
on a whitelist.

Currently whitelist is comprised of localhost, and hosts in
the repository's pgpass file. The pgpass file should include
fci and cci hosts that are safe to destroy, but shouldn't
contain fqa and cqa hosts that are not safe.

Change-Id: Ic7ecef142728bd95a00830bc739b8831fcdf1864