Added shared truststore and ant commands for updating it.
Java refuses to use system certificate storage, but there is no particular reason
we should all be prisoners of our local build environment when it comes to creating
and maintaining the local trusted-certificate list for SSL connections made from
Java code.

Created a new keystore file with the root certificates for the Wireless Generation
certificate authority and the Go Daddy CA (used by Learnosity), and added ant
targets to maintain the file, so nobody else has to read the documenation for keytool.

Change-Id: Ibc7f30bbb287165f7dd86fcd3b099d24526d449b