US23889 - added a target to create webdriver page jars, and adjusted others.
Actually, updated the existing build-test-jars target, but made it
actually do something, and run only when a specific property is set.
Only creates jars for the webdriver pages (binary and source) because
that seemed simplest.

In a similar spirit of questionable simplicity, the zip-web-assets
and prepare-web-assets-for-zip targets now only run if a particular
target is *not* set, which is probably not a great long-term plan
but will serve the current purpose adequately.

Change-Id: Ie65791f5b883282c4ec5c1c2ab9515a95515ccd8