Squashed combo commit for US23889 (don't bother merging).
US 23889 - refactored ivyTargets to be less annoying.

Created a simple "publish" macro to encapsulate the repetition between existing
(and future) ivy publication targets, and used it to replace the guts of both
ivy-publish and ivy-publish-dev.  Since this made it obvious what the problem
was with the existing way we set the revision to "dev" in the latter target,
fixed that as well.

Change-Id: I2831c74cab352c720b697969a737024b889d3e86

Made unzip-web-assets conditional on having web-assets to unzip.

Currently, projects that have no web-assets ivy directory make this target
blow up, which seems silly.  Instead, silently move on if that directory is
missing.

Change-Id: Iff75628cb8a8aac9af8039464d50af3ecd17c9f3

US23889 - made build files less insistent on standard file layout.

For the Lighthouse project, we have neither application jars nor unit,
integration, or webservice tests.  Updated the build scripts to not
fail if the directories for those are not present.

Change-Id: I391ea8ae5eaf3c0e7f21322a1ad3e00c4aa40b6e

US23889 - added a target to create webdriver page jars, and adjusted others.

Actually, updated the existing build-test-jars target, but made it
actually do something, and run only when a specific property is set.
Only creates jars for the webdriver pages (binary and source) because
that seemed simplest.

In a similar spirit of questionable simplicity, the zip-web-assets
and prepare-web-assets-for-zip targets now only run if a particular
target is *not* set, which is probably not a great long-term plan
but will serve the current purpose adequately.

Change-Id: Ie65791f5b883282c4ec5c1c2ab9515a95515ccd8

US23889 - moved some property-juggling down from projects.

Rather than repeat a lot of property-definition code in default.build.properties
and build.xml, moved it down here, with what are hoped to be adequate (if not
entirely pretty) tap-dancing to make everything work with both application and
module projects.  The appropriate publication revision for ivy should now always
be set automatically, as should the names of the standard webapp war and jar files.

Change-Id: Idf01b40203873eadde22a56f80628ada5db9116d

Added a post-build ivy-publish target, and renamed the existing dev target.

The target "ivy-publish-repeat" was not a great name, but it becomes more obvious
what the name *should* be when we create a new, low-dependency, publication target
called "ivy-publish-only".

Change-Id: Iac8a548699c0ec4aed056fba6dc3b31b60bf919c