US 23889 - refactored ivyTargets to be less annoying.
Created a simple "publish" macro to encapsulate the repetition between existing
(and future) ivy publication targets, and used it to replace the guts of both
ivy-publish and ivy-publish-dev.  Since this made it obvious what the problem
was with the existing way we set the revision to "dev" in the latter target,
fixed that as well.

Change-Id: I2831c74cab352c720b697969a737024b889d3e86