US23889 - moved some property-juggling down from projects.
Rather than repeat a lot of property-definition code in default.build.properties
and build.xml, moved it down here, with what are hoped to be adequate (if not
entirely pretty) tap-dancing to make everything work with both application and
module projects.  The appropriate publication revision for ivy should now always
be set automatically, as should the names of the standard webapp war and jar files.

Change-Id: Idf01b40203873eadde22a56f80628ada5db9116d