US23889 post-apocalypse: the Studio Problem.
Updated ivyTargets to allow projects (e.g. intel-assess) to avoid collisions between
ivysettings files, and to respect the setting of app.config.dir more completely than
it did (removing hard-coding of "conf" and "conf/base").

Updated jarTargets to use a different rule for skipping build-app-jars (and use a
shared property for the webdriver jar name), added a rule to testTargets to skip
checkstyle, and generally swept through eliminating hard-coded copies of
property-set values in as many places as possible.

Change-Id: I7c3d59c3205bf2c5180e18e6cd53781eca59ddcf