Fixer Friday: move a bunch of configuration into a function.
Repetitive (albeit clever) configuration logic in several jenkins builds is
now consolidated into a single file, which is used in all the scripts, in
hopes that we can reduce the amount of boilerplate jenkins configuration we use.

Additionally, the totally useless DB_HOSTCLASS parameter, unused since 2011, is
removed, and a few things that traditionally have the same value get defaulted
to that value instead of being set individually.

Change-Id: I0f1a46040fd5df18bb5887bfcfba1e109356bab9