Merge branches should always merge to correct branch
We switched Jenkins' settings to set the base branch for
this build as the release branch, so the Recent Changes
view showed changes that were on the release branch,
instead of changes that wereon the master branch.

After this, the script was merging release into release,
instead of release into master. So, now it'll always
merge the head branch into the base branch, regardless
of where it starts from.

Also, adding debugging output to the script.

Change-Id: I42dd913b95aea8600932a86c2fe1812abf1774a7