Updated default dependencies.
Added a "stable" branch property for things where Beacon does not need
latest/greatest but does need stability, and made common-api-java,
mas-java-client (which now has branches!) and java-logging use it.

Updated the actual dependency revision for mas-java-client to a reasonable
range based on current PSG development practices, and common-api-java and
core-lib-java to ranges that catch currently-available things.

Using this properties update will require an update to the ivy configuration
for each project, because of the introduction of branching to mas-java-client.

Change-Id: I3ce1fc775347666efbcab5e12da0ca9f68f41c12